Split tests into two classes.
