+++ src/test/java/com/ledger/u2f/U2FTest.java
@@ -289,29 +289,7 @@ public class U2FTest extends SimulatorTestBase {
         assertThat(getDataAPDU.getSW(), is(ISO7816.SW_CONDITIONS_OF_USE_NOT_SATISFIED));
     }
     
-    @Test
-    public void testCompareConstantTimeEmpty() {
-        byte[] array1 = new byte[]{};
-        byte[] array2 = new byte[]{0x01, 0x01, 0x01};
-        short zero = (short)0;
-        assertThat(FIDOUtils.compareConstantTime(array1, zero, array2, zero, zero), is(false));
-    }
-    
-    @Test
-    public void testCompareConstantTimeNonEqual() {
-        byte[] array1 = new byte[]{0x01, 0x02, 0x01};
-        byte[] array2 = new byte[]{0x01, 0x01, 0x01};
-        short zero = (short)0;
-        assertThat(FIDOUtils.compareConstantTime(array1, zero, array2, zero, (short)3), is(false));
-    }
-    
-    @Test
-    public void testCompareConstantTimeEqual() {
-        byte[] array1 = new byte[]{0x0F, 0x01, 0x02, 0x03, 0x0F};
-        byte[] array2 = new byte[]{0x00, 0x00, 0x00, 0x01, 0x02, 0x03, 0x00, 0x00};
-        assertThat(FIDOUtils.compareConstantTime(array1, (short)1, array2, (short)3, (short)3), is(true));
-    }
-    
+
     @Test
     public void testSetAttestationCertWrongLength() {
         prepareApplet(INSTALL_FLAG_ENABLE_USER_PRESENCE, attestationCert.length - 8, attestatioPrivkey);
